+++ reduce_src/PDBrec.h
@@ -218,8 +218,8 @@ public:
       return n;
       //if (errmsg) throw std::runtime_error(errmsg);
    }
-   float occupancy()      const { return _rep->_r.atom.occupancy;          }
-   float tempFactor()     const { return _rep->_r.atom.tempFactor;         }
+   float occupancy()      const { return static_cast<float>(_rep->_r.atom.occupancy);   }
+   float tempFactor()     const { return static_cast<float>(_rep->_r.atom.tempFactor);  }
    const char*segidLabel() const{ return _rep->_r.atom.segID;              }
    const char*elemLabel()  const{ return _rep->_r.atom.element;            }
    const char*chargeLabel()const{ return _rep->_r.atom.charge;             }
