+++ reduce_src/RotMethyl.cpp
@@ -212,7 +212,7 @@ double RotMethyl::scoreThisAngle(AtomPositions &xyz, DotSphManager& dotBucket,
 		const  std::shared_ptr<PDBrec> thisAtom = *alst;
 
 		double val = xyz.atomScore(*thisAtom, thisAtom->loc(),
-			thisAtom->vdwRad() + probeRadius + maxVDWrad,
+      static_cast<float>(thisAtom->vdwRad() + probeRadius + maxVDWrad),
 			//apl procrastinate nearby list computation until AtomPositions decides to score
 			*(_bnded[i]), dotBucket.fetch(thisAtom->vdwRad()), probeRadius, FALSE,
 			bumpSubScore, hbSubScore, subBadBump);
